Added permissions checks to API and added DELETE for folders API
on_delete on FTLDocument have been switched to PROTECT. This will
enforce programmatic deletion of all documents when org, user or folder
are removed from DB. Also ensure that documents binary are also deleted.

New FTLModelPermissions class for API view. It adds instance permission
check to every API call. The permmissions are handled by Django
standard permission backend.

Added additional `org` checks to API calls.

Added ATOMIC_REQUESTS to settings. Upon reading MPTT (folders manager)
docs about concurrent requests, it appears that MPTT does not lock or
use transaction by default which can results to invalid folders state in
DB.

Added basic permissions to admin creation and user creation form.

+++ ftl/core/views.py
@@ -2,14 +2,15 @@ import json
 import os
 
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseForbidden
 from django.shortcuts import render, get_object_or_404
 from django.views import View
 from rest_framework import generics, views
+from rest_framework.exceptions import ValidationError
 from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
 
-from core.models import FTLDocument, FTLFolder
+from core.models import FTLDocument, FTLFolder, FTLModelPermissions
 from core.serializers import FTLDocumentSerializer, FTLFolderSerializer
 
 
@@ -26,14 +27,19 @@ class DownloadView(View):
     http_method_names = ['get']
 
     def get(self, request, *args, **kwargs):
+        # We don't use FTLModelPermissions because it only works with APIView
+        if not self.request.user.has_perm('core.view_ftldocument'):
+            raise HttpResponseForbidden
+
         doc = get_object_or_404(FTLDocument.objects.filter(ftl_user=self.request.user, pid=kwargs['uuid']))
         response = HttpResponse(doc.binary, 'application/octet')
         response['Content-Disposition'] = 'attachment; filename="%s"' % doc.binary.name
         return response
 
 
-class FTLDocumentList(generics.ListCreateAPIView):
+class FTLDocumentList(generics.ListAPIView):
     serializer_class = FTLDocumentSerializer
+    permission_classes = (FTLModelPermissions,)
 
     def get_queryset(self):
         current_folder = self.request.query_params.get('level', None)
@@ -47,13 +53,11 @@ class FTLDocumentList(generics.ListCreateAPIView):
 
         return queryset
 
-    def perform_create(self, serializer):
-        serializer.save()  # TODO Do we need this?
-
 
 class FTLDocumentDetail(generics.RetrieveUpdateDestroyAPIView):
     serializer_class = FTLDocumentSerializer
     lookup_field = 'pid'
+    permission_classes = (FTLModelPermissions,)
 
     def get_queryset(self):
         return FTLDocument.objects.filter(ftl_user=self.request.user)
@@ -65,19 +69,25 @@ class FTLDocumentDetail(generics.RetrieveUpdateDestroyAPIView):
         serializer.save(ftl_user=self.request.user)
 
     def perform_destroy(self, instance):
-        binary = instance.binary
-        super().perform_destroy(instance)
-        binary.file.close()
-        os.remove(binary.file.name)
+        if instance.org == self.request.user.org:
+            binary = instance.binary
+            super().perform_destroy(instance)
+            binary.file.close()
+            os.remove(binary.file.name)
+        else:
+            raise ValidationError("Trying to delete document from wrong user!")
 
 
 class FileUploadView(views.APIView):
     parser_classes = (MultiPartParser,)
     serializer_class = FTLDocumentSerializer
+    permission_classes = (FTLModelPermissions,)
+    # Needed for applying permission checking on view that don't have any queryset
+    queryset = FTLDocument.objects.none()
 
     def post(self, request):
         file_obj = request.data['file']
-        payload = json.loads(request.data['json'])  # Nothing for now
+        payload = json.loads(request.data['json'])
 
         # TODO check for empty form
 
@@ -101,6 +111,7 @@ class FileUploadView(views.APIView):
 class FTLFolderList(generics.ListCreateAPIView):
     serializer_class = FTLFolderSerializer
     pagination_class = None
+    permission_classes = (FTLModelPermissions,)
 
     def get_queryset(self):
         current_folder = self.request.query_params.get('level')
@@ -115,3 +126,21 @@ class FTLFolderList(generics.ListCreateAPIView):
 
     def perform_create(self, serializer):
         serializer.save(org=self.request.user.org)
+
+
+class FTLFolderDetail(generics.RetrieveUpdateDestroyAPIView):
+    serializer_class = FTLFolderSerializer
+    lookup_field = 'id'
+    permission_classes = (FTLModelPermissions,)
+
+    def get_queryset(self):
+        return FTLFolder.objects.filter(org=self.request.user.org)
+
+    def perform_update(self, serializer):
+        serializer.save(org=self.request.user.org)
+
+    def perform_destroy(self, instance):
+        if instance.org == self.request.user.org:
+            instance.delete()
+        else:
+            raise ValidationError('Trying to delete a folder from wrong user!')
